fix(ui): 11 bug fixes and UI improvements
- fix(addons): correct badge color typo bg0 -> bg for PARTS/LABOR categories
- fix(reports): resolve photo storage paths to signed URLs and fix gallery key flicker
- fix(ui): improve detail button text contrast for all themes
- refactor(addons): hide stock columns and inputs from catalog UI (keep DB data)
- feat(technician): add decimal qty support (DB migration + UI)
- fix(ui): add visible background to inactive tab triggers
- fix(ui): remove janky tab animations, use static active indicator, strip dead code
- fix(ac-units): fix edit form not populating existing AC data
- fix(ac-units): change edit/delete buttons to icon-only with proper touch targets
- feat(technician): add barcode scanner for AC serial number input
- chore: fix type and lint errors from UI fixes batch

BREAKING: requires DB migration (supabase/migrations/10_decimal_qty.sql)

+++ supabase/migrations/20260616_technician_submit_report_v3.sql
@@ -46,7 +46,7 @@ BEGIN
 
       IF v_catalog_price IS NOT NULL THEN
         v_material := jsonb_set(v_material, '{unit_price}', to_jsonb(v_catalog_price));
-        v_material := jsonb_set(v_material, '{total}', to_jsonb(v_catalog_price * COALESCE((v_material->>'qty')::int, 1)));
+        v_material := jsonb_set(v_material, '{total}', to_jsonb(v_catalog_price * COALESCE((v_material->>'qty')::numeric, 1.0)));
       END IF;
     END IF;
 
@@ -188,7 +188,7 @@ BEGIN
 
     IF v_addon_id IS NOT NULL AND COALESCE((v_material->>'is_manual')::boolean, false) = false THEN
       UPDATE addon_catalog
-      SET stock_quantity = stock_quantity - COALESCE((v_material->>'qty')::int, 1),
+      SET stock_quantity = stock_quantity - COALESCE((v_material->>'qty')::numeric, 1.0),
           updated_at = NOW()
       WHERE addon_id = v_addon_id;
     END IF;
